Extract image row to reusable component
